feat(set-current): add subcommand
As said on reddit by many user, it might be useful to save keys, mail and user as a per repo basis.
With this feature one can `saint-peter set-current <key>` to set it once and saint-peter will handle
the next commits and push.
If the keys, use or mail needs to be updated, one can update them in the configuration and run the above command
and it will handle all the modifications.